requirements.txt
pysimplegui<5
python-vlc
requests
